+++ examples/chemical-potential.py
@@ -74,7 +74,7 @@ import maicos
 # input instead of the Volume.
 
 
-def mu(rho, T, m):
+def mu(rho: np.ndarray, T: float, m: float) -> np.ndarray:
     """Calculate the chemical potential.
 
     The chemical potential is calculated from the density: mu = R T log(rho. / m)
@@ -111,7 +111,7 @@ def mu(rho, T, m):
 # The implemented function looks like this.
 
 
-def dmu(rho, drho, T):
+def dmu(rho: np.ndarray, drho: np.ndarray, T: float) -> np.ndarray:
     """Calculate the error of the chemical potential.
 
     The error is calculated from the density using propagation of uncertainty.